Moved tool functions from radiation_profile.py to radiation_tools.py
